requirements.txt
accelerate==0.16.0
bitsandbytes==0.37.0
bitsandbytes_cuda117==0.30.1
datasets==2.10.0
diffusers==0.14.0
gradio==3.20.1
huggingface_hub==0.12.1
liteobj==0.0.2
numpy==1.23.3
omegaconf==2.2.3
packaging==23.0
Pillow==9.4.0
python-dotenv==1.0.0
pytorch_lightning==1.9.4
taming_transformers==0.0.1
torch==1.13.1
torchaudio==0.13.0
torchvision==0.12.0+cu113
tqdm==4.64.0
transformers==4.26.1
transformers.egg==info
xformers==0.0.17.dev466
